requirements.txt
alembic==1.0.11
aniso8601==6.0.0
astroid==2.2.5
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.8
Flask-GraphQL==2.0.0
Flask-GraphQL-Auth==1.2.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
graphene==2.1.6
graphene-sqlalchemy==2.2.1
graphql-core==2.2
graphql-relay==0.4.5
graphql-server-core==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
Mako==1.0.12
MarkupSafe==1.1.1
mccabe==0.6.1
promise==2.2.1
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.5
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.2
